+++ src/lib.rs
@@ -58,6 +58,7 @@ extern crate nom;
 
 pub mod oath;
 pub mod pass;
+pub mod key;
 
 #[cfg(feature = "cbindings")]
 extern crate libc;