+++ tests/libreauth_tests.h
@@ -39,3 +39,4 @@ void        test_name(const char *name);
 uint32_t    test_hotp(void);
 uint32_t    test_totp(void);
 uint32_t    test_pass(void);
+uint32_t    test_key(void);