+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [Unreleased][unreleased]
+
+### Added
+- Random key generation module.
+
 ## [0.6.1] - 2017-12-17
 
 ### Changed